implement offset tracking
Offset tracking is implemented as two new chunk types, one for snapshots
that are written every time the writer opens a new segment and one for
delta updates that are written in real time as they come in. The writer
maintains a map of tracking ids to offsets and can be queried for a
given tracking id.