Offset readers handle missing segments
When retention "takes-over" an offset reader it is possible that a
segment later than the one currently being read is deleted. When the
offset reader reaches the end of it's current segment and finds the next
segment missing it gets stuck and things it has reached the end of the
stream. This change expands the use of the atomic value for the commit
index to also include the first index in the stream which the reader can
query to ensure it correctly can work out what the next available
segment is.