Fix out of order timestamps of tracking snapshots
Because tracking snapshots are written before the first user chunk in each
segment and the timestamps of the snapshot would be generated _after_ the
user chunk the could end up out of order. This change passes the user chunk
timestamp to the tracking snapshot function so that the same timestamp is used.