Add max_segment_size_chunks
We've observed performance degradation when there is lots of chunks in a
single segment. That's because the segment and/or its index need to be
scanned in certain situations and reading chunk by chunk is slow.

For now, it's compile-time setting with a default of 256000. We may
change that into a service configuration property in the future, if
needed.