Fix bug when truncating replica stream with no overlap
When a replica is initialising and there is local data but none of
it overlaps with the writer's source data the local stream was not
truncated resulting in an out of order error when the replica received
it's first chunk. This change ensures the entire log is deleted when
there is not overlap.