Restart replica TLS listener in case of handshake error
An handshake error in this case is most likely that the client/reader
passes an unexpected hostname and not the one in the server
certificate. The replica then restart the TLS listener and the client
can retry another host, that will hopefully be compliant with
what the server certificate presents.