Avoid race condition when attaching at last + 1
When an offset reader wants to attach at a given offset and this offset
is the next offset to be written in the stream there was a race condition
when we first detect the offset is out of range then retry attach with the
`next` offset spec. if a chunk was written in between these two events we
could end up skipping one or more chunks of data.