Make init_reader and init_data_reader use a consistent return value
As discovered in #68, there is a typo in one of the returned
error atoms.

We stick to "invalid_last_offset_epoch" over "invalid_last_epoch_offset"
because the check that detects an invalid value
in osiris_log:check_chunk_has_expected_epoch/3 checks for epoch
validity, not offset validity.

This is a potentially breaking change but even in RabbitMQ-related
projects, this specific error was not handled.