Add hook for file:consult/1 for TLS replication
file:consult/1 is OK to parse a configuration file with some
TLS settings, but it does not support parsing "complex"
settings like a function to validate the hostname.

This commit introduces a hook so that the calling application
can provide a custom function instead of the default file:consult/1
and this way easily support parsing the app settings.

Fixes #78