Prevent crash when init_offset_reader in empty dir
Prior to this commit, there was a badmatch crash
when init_offset_reader was called for a directory that does not contain
any index (and segment) files.

Let's instead return an {error, no_index_file} which can be better
handled by applications of this library.

(In a past version of osiris exit(no_segments_found) was thrown.)