Share last chunk id with readers to avoid unnecessary work
When a reader reaches end_of_stream it tries to open the next potential segment file.

Most of the time there wont be one and this work has a degree of overhead.

This commit adds another element to the shared atomic value that is passed to data and
offset readers so that they can check this first to see if the local reader thinks there should be data in the log.