Fix diverged replica truncation bug
When a former writer becomes replica and has uncommitted entriees in
it's and it goes to truncate the log based on a new leader's tail and this
new leader has a last chunk_id in a new epoch that is exactly the last
chunk_id + the number of entries (i.e. next chunk id) of the diverged leader's
log it would skip truncation and later error when the replica reader
wanted to attach.