Chunk filtering
This implements a reasonable efficient way to avoid reading chunks
that do not contain messages that the reader is not interested in.

For each chunk where there is at least one write that includes a
filter value a bloom filter will be added to the chunk immediately
after the chunk header. The chunk filter is calculated from all the
filter values in the chunk. If there are writes that contain no
filter value the chunk filters first bit will also be set to indicate
there are "unfiltered" entries in the chunk.

At read time an offset reader can supply one or more filter values that are used
to match against the bloom filter. If none of the read filter values
matches the chunk body is not read and the reader moves on to the next
chunk and so on.

There is an option also to include "unfiltered" entries in the match
evaluation.

The bloom filter can be in the size range of 16-255 bytes with the default
the smallest value at 16 bytes. This should be good for distinguishing
between ~50 unique filter values.

The bloom filter size is recorded in the first of the spare 4 bytes
at the end of the chunk header.