Fix offset reader race condition
Where an offset reader on a replica node can try to read a chunk
before it has been fully written and cause a crc failure.

This could occur if the replica updated the committed offset before
the corresponding chunk has been written. The reader would only check
the committed offset before attempting to read a chunk. If partial data was
available it would not return end_of_stream as expected but instead
fail when the full data section could not be read.

The fix is to check both the committed offset and the last chunk id
before attempting a read.