Handle one more case of corrupt files at startup
After RabbitMQ was shut down abruptily the last index file did not have
a sequence file. In this state the stream cannot start up and crashes in
a cicle.

```
initial call: osiris_writer:init/1
...
reason: {{case_clause,missing_file},
         [{gen_batch_server,handle_continue,2,
                            [{file,"src/gen_batch_server.erl"},
                             {line,413}]},
```

It was found that the `missing_file` is thrown from
`osiris_log:maybe_fix_corrupted_files/1`

In this case just delete the index file as no data of the missing
segment file can be recovered. (same as if the segment file would exist
but withou any chunks)