Be more defensive when evaluating retention.
It is possible that osiris_log:evaluate_retention/1 may throw an
error, in particular `missing_file` which could occur if an osiris
member was deleted whilst an evaluation request was outstanding.
 Currently this crashes the evaluation process which isn't a big deal
as it will restart but we'd lose any pending evaluation requests when this
happen so it is better to try to avoid that.