Improve logging during replicat initialization
Starting a replica and its reader can temporarily in some cases and
environment (e.g. when restarting a node on K8S, when DNS is not sorted
out yet). The current logging messages can be misleading and verbose.

This commit tries to avoid crash reports when a replica reader cannot
connect to its replica. This is usually a transient issue and the whole
crash report does not help much and clutters the log file.

References #186