Add osiris_log:overview/1 to the manifest behaviour
This change also refactors the overview type from being

    {osiris_log:range(), [{osiris:epoch(), osiris:offset()}]}

to being a map where both of those are required:

    #{range := osiris_log:range(),
      epoch_offsets := [{osiris:epoch(), osiris:offset()}]}

This allows other manifest implementations to stick other data in the
map. This map is passed through to replicas in `init_acceptor/2` and
`acceptor_manifest/2`, so they can use information from the manifest
from the writer's node.