Add a `{'fun', fun()}` retention spec
This can be used flexibly to evaluate retention depending on the name or
contents of index files. You could pass in a function which would return
true until the offset in the index file's name exceeds some given offset
for example, as a way to truncate everything up to that offset or to
guarantee that the offset (for example an uncommitted one) won't be
truncated.