Replace `osiris_log:overview/1` tuple with a map
This is important for allowing other manifest implementations (see child
commits) to include additional info in the map. The range and epoch
offsets are the only required keys and other manifests might include
other info.

While this change is not complicated, it is breaking and needs backward
compatibility clauses in callers of `overview/1` and `init_acceptor/3`.
And it ends up moving a lot of code around, inflating the diff.