osiris_writer: Track replica staleness and backlog
This introduces a metric calculated at every batch which records the
difference between the timestamps of the last chunks in the logs of the
writer and its replicas, and the sum of offsets that need to be
replicated. This can be used to watch replicas catch up on replication,
or to diagnose situations when replication is being starved out
(network-wise) by high-throughput publishing. The replication diff
should usually be low but, for streams seeing traffic, non-zero.

This calculation is also used in the stream coordinator when adding a
member, via `osiris_writer:query_replication_state/1`. With this change
the stream coordinator could be updated to use the counter instead of
calling the writer.